+++ script/provider/provider.lua
@@ -652,9 +652,6 @@ proto.on('textDocument/semanticTokens/full', function (params)
         text  = files.getText(uri)
     end
     local results = core(uri, 0, #text)
-    if not results or #results == 0 then
-        return nil
-    end
     return {
         data = results
     }
@@ -674,9 +671,6 @@ proto.on('textDocument/semanticTokens/range', function (params)
     local start  = define.offset(lines, text, params.range.start)
     local finish = define.offset(lines, text, params.range['end'])
     local results = core(uri, start, finish)
-    if not results or #results == 0 then
-        return nil
-    end
     return {
         data = results
     }